# api/openapi.yaml
openapi: 3.0.3
info:
  title: JapanDigitalPostService API
  version: 0.1.0
  description: >
    日本邮政地址检索服务。基于 token 认证，提供邮编/都道府県/市区町村/模糊查询，
    以及同步状态与 token 管理。本契约为 spec-first 源文件，用 oapi-codegen 生成 Go server。
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /addresses:
    get:
      summary: 地址查询（邮编/都道府県/市区町村/模糊）
      operationId: searchAddresses
      parameters:
        - { name: zipcode, in: query, schema: { type: string }, description: 7 位邮编或前缀 }
        - { name: prefecture, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string }, description: 跨字段模糊关键字 }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 20 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: 查询结果（含 too_many_results / timeout 等状态）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResult' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '504': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Address:
      type: object
      properties:
        zipcode: { type: string }
        jis_code: { type: string }
        prefecture: { type: string }
        prefecture_kana: { type: string }
        city: { type: string }
        city_kana: { type: string }
        town: { type: string }
        town_kana: { type: string }
    SearchResult:
      type: object
      description: >
        查询成功响应。超时/参数非法/未认证等不返回本结构，而是 Error（见各端点的
        400/401/404/504 响应）。
      required: [status, total_count, returned_count, truncated, items]
      properties:
        status:
          type: string
          description: ok = 正常或零命中；too_many_results = 命中数超过 MAX_TOTAL。
          enum: [ok, too_many_results]
        total_count:
          type: integer
          description: 满足条件的总命中数（可能远大于 returned_count）。
        returned_count:
          type: integer
          description: 本次实际返回的 items 条数（模糊查询硬上限 = FUZZY_LIMIT，默认 20）。
        truncated:
          type: boolean
          description: total_count > returned_count 时为 true。
        address_count:
          type: integer
          description: >
            仅 GET /addresses/{zipcode} 返回：该邮编定位到的地址条数
            （一个邮编可对应多个町域）。列表/模糊查询不含此字段。
        items:
          type: array
          items: { $ref: '#/components/schemas/Address' }
...
    Error:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: 机器可读状态码（spec §7）。
          enum: [too_many_results, timeout, invalid_request, unauthorized, forbidden, not_found, rate_limited, internal_error, sync_running, unsupported_file, unzip_failed, csv_format_error, import_failed]
        message: { type: string, description: 人类可读说明 }
        trace_id: { type: string, description: 关联请求的 X-Request-Id，便于排查 }